% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENVA_RBD.R
\name{GENVA_RBD}
\alias{GENVA_RBD}
\title{Provides RBD analysis and Path analysis}
\usage{
GENVA_RBD(DATA)
}
\arguments{
\item{DATA}{Experimental data from RBD}
}
\description{
Provides RBD analysis and Path analysis
}
