% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENVA_RBD.R
\name{GENVA_RBD}
\alias{GENVA_RBD}
\title{Provides RBD analysis and path analysis}
\usage{
GENVA_RBD(DATA)
}
\arguments{
\item{DATA}{RBD experiment data}
}
\description{
Provides RBD analysis and path analysis
}
